Pull `is_range_literal` out into `lowering`
